Moved ledger state operations into a separate file.
