+++ ledger/ledger-on-memory/src/main/scala/com/daml/ledger/on/memory/InMemoryLedgerStateOperations.scala
@@ -1,0 +1,40 @@
+package com.daml.ledger.on.memory
+
+import com.daml.ledger.on.memory.InMemoryLedgerReaderWriter.Index
+import com.daml.ledger.on.memory.InMemoryState.MutableLog
+import com.daml.ledger.participant.state.kvutils.KVOffset
+import com.daml.ledger.participant.state.kvutils.api.LedgerRecord
+import com.daml.ledger.participant.state.v1.Offset
+import com.daml.ledger.validator.BatchingLedgerStateOperations
+import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
+
+import scala.concurrent.{ExecutionContext, Future}
+
+private[memory] final class InMemoryLedgerStateOperations(
+    log: InMemoryState.MutableLog,
+    state: InMemoryState.MutableState,
+)(implicit executionContext: ExecutionContext)
+    extends BatchingLedgerStateOperations[Index] {
+  import InMemoryLedgerStateOperations.appendEntry
+
+  override def readState(keys: Seq[Key]): Future[Seq[Option[Value]]] =
+    Future.successful(keys.map(state.get))
+
+  override def writeState(keyValuePairs: Seq[(Key, Value)]): Future[Unit] = {
+    state ++= keyValuePairs
+    Future.unit
+  }
+
+  override def appendToLog(key: Key, value: Value): Future[Index] =
+    Future.successful(appendEntry(log, LedgerRecord(_, key, value)))
+}
+
+object InMemoryLedgerStateOperations {
+  private[memory] def appendEntry(log: MutableLog, createEntry: Offset => LedgerRecord): Index = {
+    val entryAtIndex = log.size
+    val offset = KVOffset.fromLong(entryAtIndex.toLong)
+    val entry = createEntry(offset)
+    log += entry
+    entryAtIndex
+  }
+}
